+++ src/main/java/open/commons/spring/web/authority/metadata/AuthorizedMetadataBuilder.java
@@ -43,6 +43,7 @@ import org.springframework.context.annotation.Bean;
 import open.commons.core.utils.ExceptionUtils;
 import open.commons.core.utils.StringUtils;
 import open.commons.spring.web.authority.AuthorizedField;
+import open.commons.spring.web.authority.AuthorizedObject;
 import open.commons.spring.web.beans.authority.IFieldAccessAuthorityProvider;
 import open.commons.spring.web.beans.authority.IUnauthorizedFieldHandler;
 import open.commons.spring.web.servlet.InternalServerException;
@@ -68,7 +69,43 @@ public class AuthorizedMetadataBuilder {
      * Stage 1: 시작점
      */
     public interface Builder {
+
+        /**
+         * 1개의 {@link AuthorizedObjectMetadata}를 생성하는 빌더를 제공합니다. <br>
+         * 
+         * <pre>
+         * [개정이력]
+         *      날짜    	| 작성자	|	내용
+         * ------------------------------------------
+         * 2025. 6. 20.		박준홍			최초 작성
+         * </pre>
+         *
+         * @return
+         *
+         * @since 2025. 6. 18.
+         * @version 0.8.0
+         * @author Park, Jun-Hong parkjunhong77@gmail.com
+         */
         ObjectBuilder object();
+
+        /**
+         * 여러 개의 {@link AuthorizedObjectMetadata}를 생성하는 빌더를 제공합니다. <br>
+         * <br>
+         * 
+         * <pre>
+         * [개정이력]
+         *      날짜    	| 작성자	|	내용
+         * ------------------------------------------
+         * 2025. 6. 20.		박준홍			최초 작성
+         * </pre>
+         *
+         * @return
+         *
+         * @since 2025. 6. 20.
+         * @version 0.8.0
+         * @author Park, Jun-Hong parkjunhong77@gmail.com
+         */
+        ObjectsBuilder objects();
     }
 
     /**
@@ -79,6 +116,11 @@ public class AuthorizedMetadataBuilder {
         @Override
         public ObjectBuilder object() {
             return new ObjectBuilderImpl();
+        }
+
+        @Override
+        public ObjectsBuilder objects() {
+            return new ObjectsBuilderImpl();
         }
 
         private static boolean allowAccess(AccessibleObject o) {
@@ -95,11 +137,35 @@ public class AuthorizedMetadataBuilder {
          * [개정이력]
          *      날짜    	| 작성자	|	내용
          * ------------------------------------------
-         * 2025. 6. 19.		박준홍			최초 작성
+         * 2025. 6. 20.		박준홍			최초 작성
          * </pre>
          *
+         * @param targetClass
          * @param builder
-         * @param target
+         *
+         * @since 2025. 6. 20.
+         * @version 0.8.0
+         * @author Park, Jun-Hong parkjunhong77@gmail.com
+         */
+        private static <T> T newObject(Class<T> targetClass, Object builder) {
+            return newObject(targetClass, builder, null);
+        }
+
+        /**
+         * 
+         * <br>
+         * 
+         * <pre>
+         * [개정이력]
+         *      날짜    	| 작성자	|	내용
+         * ------------------------------------------
+         * 2025. 6. 19.		박준홍			최초 작성
+         * </pre>
+         * 
+         * @param targetClass
+         *            생성할 객체
+         * @param builder
+         *            빌더 객체
          * @param postprocessors
          *            <li>key: 필드 이름
          *            <li>value: 후처리 모듈
@@ -108,8 +174,8 @@ public class AuthorizedMetadataBuilder {
          * @version 0.8.0
          * @author Park, Jun-Hong parkjunhong77@gmail.com
          */
-        private static void set(Object builder, Object target, Map<String, Function<Object, Object>> postprocessors) {
-            Class<?> targetClass = target.getClass();
+        private static <T> T newObject(Class<T> targetClass, Object builder, Map<String, Function<Object, Object>> postprocessors) {
+            T newObject = null;
             Boolean targetFieldAccessible = null;
             String targetFieldName = null;
             Method targetMethod = null;
@@ -118,45 +184,50 @@ public class AuthorizedMetadataBuilder {
             Field builderField = null;
             Boolean builderFieldAccessible = null;
             Object value = null;
-            for (Field targetField : targetClass.getDeclaredFields()) {
-                try {
-                    // #1. 대상 클래스 필드 접근 허용
-                    targetFieldAccessible = allowAccess(targetField);
-                    // #2. builder 객체에서 동일한 이름의 필드 정보 조회
-                    builderField = ClassInspector.getDeclaredFieldIfExist(builderClass, targetFieldName = targetField.getName());
-                    if (builderField == null) {
-                        continue;
-                    }
+            try {
+                newObject = targetClass.newInstance();
+                for (Field targetField : targetClass.getDeclaredFields()) {
+                    try {
+                        // #1. 대상 클래스 필드 접근 허용
+                        targetFieldAccessible = allowAccess(targetField);
+                        // #2. builder 객체에서 동일한 이름의 필드 정보 조회
+                        builderField = ClassInspector.getDeclaredFieldIfExist(builderClass, targetFieldName = targetField.getName());
+                        if (builderField == null) {
+                            continue;
+                        }
 
-                    builderFieldAccessible = allowAccess(builderField);
-                    value = builderField.get(builder);
-                    // builderField.get(builder) 후처리
-                    if (postprocessors != null && postprocessors.containsKey(targetFieldName)) {
-                        value = postprocessors.get(builderField.getName()).apply(value);
-                    }
+                        builderFieldAccessible = allowAccess(builderField);
+                        value = builderField.get(builder);
+                        // builderField.get(builder) 후처리
+                        if (postprocessors != null && postprocessors.containsKey(targetFieldName)) {
+                            value = postprocessors.get(builderField.getName()).apply(value);
+                        }
 
-                    if (value == null) {
-                        continue;
-                    }
+                        if (value == null) {
+                            continue;
+                        }
 
-                    targetMethod = targetClass.getMethod(String.join("", "set", StringUtils.toUpperCase(targetFieldName, 0)), targetField.getType());
-                    targetMethod.invoke(target, value);
-                } catch (Exception e) {
-                    throw ExceptionUtils.newException(InternalServerException.class, e,
-                            "데이터 처리 도중 오류가 발생하였습니다. target.class=%s, target.method=%s, target.object=%s / builder.class=%s, builder.field=%s, builder.object=%s" //
-                            , targetClass, targetMethod, target, builderClass, builderField, builder);
-                } finally {
-                    if (targetFieldAccessible != null && targetField != null) {
-                        targetField.setAccessible(targetFieldAccessible);
-                    }
-                    if (builderFieldAccessible != null && builderField != null) {
-                        builderField.setAccessible(builderFieldAccessible);
-                    }
+                        targetMethod = targetClass.getMethod(String.join("", "set", StringUtils.toUpperCase(targetFieldName, 0)), targetField.getType());
+                        targetMethod.invoke(newObject, value);
+                    } finally {
+                        if (targetFieldAccessible != null && targetField != null) {
+                            targetField.setAccessible(targetFieldAccessible);
+                        }
+                        if (builderFieldAccessible != null && builderField != null) {
+                            builderField.setAccessible(builderFieldAccessible);
+                        }
 
-                    targetFieldAccessible = null;
-                    builderFieldAccessible = null;
+                        targetFieldAccessible = null;
+                        builderFieldAccessible = null;
+                    }
                 }
+            } catch (Exception e) {
+                throw ExceptionUtils.newException(InternalServerException.class, e,
+                        "데이터 처리 도중 오류가 발생하였습니다. target.class=%s, target.method=%s, target.object=%s / builder.class=%s, builder.field=%s, builder.object=%s" //
+                        , targetClass, targetMethod, newObject, builderClass, builderField, builder);
             }
+
+            return newObject;
         }
 
         /**
@@ -174,9 +245,7 @@ public class AuthorizedMetadataBuilder {
             }
 
             public AuthorizedFieldMetadata build() {
-                AuthorizedFieldMetadata f = new AuthorizedFieldMetadata();
-                set(this, f, null);
-                return f;
+                return newObject(AuthorizedFieldMetadata.class, this);
             }
 
             public void fieldHandleBean(String fieldHandleBean) {
@@ -199,7 +268,7 @@ public class AuthorizedMetadataBuilder {
         private static class AuthorizedObjectMetadataBuilder {
             static Map<String, Function<Object, Object>> pp = new HashMap<>();
             static {
-                pp.put("fields", o -> ((List<FieldBuilder>) o).stream().map(b -> b.build()).collect(Collectors.toList()));
+                pp.put("objects", o -> ((List<FieldBuilder>) o).stream().map(b -> b.build()).collect(Collectors.toList()));
             }
             private String authorityBean;
             private String fieldHandleBean;
@@ -211,9 +280,7 @@ public class AuthorizedMetadataBuilder {
             }
 
             public AuthorizedObjectMetadata build() {
-                AuthorizedObjectMetadata o = new AuthorizedObjectMetadata();
-                set(this, o, pp);
-                return o;
+                return newObject(AuthorizedObjectMetadata.class, this, pp);
             }
 
             public void fieldHandleBean(String fieldHandleBean) {
@@ -234,37 +301,37 @@ public class AuthorizedMetadataBuilder {
          */
         private static class FieldBuilderImpl implements FieldBuilder {
 
-            private final AuthorizedFieldMetadataBuilder fieldBuilder = new AuthorizedFieldMetadataBuilder();
+            private final AuthorizedFieldMetadataBuilder afmBuilder = new AuthorizedFieldMetadataBuilder();
 
             public FieldBuilderImpl() {
             }
 
             @Override
             public FieldBuilder authorityBean(String authorityBean) {
-                fieldBuilder.authorityBean(authorityBean);
+                afmBuilder.authorityBean(authorityBean);
                 return this;
             }
 
             @Override
             public AuthorizedFieldMetadata build() {
-                return fieldBuilder.build();
+                return afmBuilder.build();
             }
 
             @Override
             public FieldBuilder fieldHandleBean(String fieldHandleBean) {
-                fieldBuilder.fieldHandleBean(fieldHandleBean);
+                afmBuilder.fieldHandleBean(fieldHandleBean);
                 return this;
             }
 
             @Override
             public FieldBuilder handleType(int handleType) {
-                fieldBuilder.handleType(handleType);
+                afmBuilder.handleType(handleType);
                 return this;
             }
 
             @Override
             public FieldBuilder name(String name) {
-                fieldBuilder.name(name);
+                afmBuilder.name(name);
                 return this;
             }
         }
@@ -274,38 +341,71 @@ public class AuthorizedMetadataBuilder {
          */
         private static class ObjectBuilderImpl implements ObjectBuilder {
 
-            private final AuthorizedObjectMetadataBuilder objectBuilder = new AuthorizedObjectMetadataBuilder();
+            private final AuthorizedObjectMetadataBuilder abmBuilder = new AuthorizedObjectMetadataBuilder();
             private final List<FieldBuilder> fields = new ArrayList<>();
 
             @Override
             public ObjectBuilder authorityBean(String authorityBean) {
-                objectBuilder.authorityBean(authorityBean);
+                abmBuilder.authorityBean(authorityBean);
                 return this;
             }
 
             @Override
             public AuthorizedObjectMetadata build() {
-                objectBuilder.fields(fields);
-                return objectBuilder.build();
+                abmBuilder.fields(fields);
+                return abmBuilder.build();
             }
 
             @SuppressWarnings("unused")
             @Override
             public ObjectBuilder field(Function<FieldBuilder, FieldBuilder> consumer) {
-                FieldBuilderImpl fb = new FieldBuilderImpl();
-                fields.add(consumer.apply(fb));
+                FieldBuilderImpl builder = new FieldBuilderImpl();
+                fields.add(consumer.apply(builder));
                 return this;
             }
 
             @Override
             public ObjectBuilder fieldHandleBean(String fieldHandleBean) {
-                objectBuilder.fieldHandleBean(fieldHandleBean);
+                abmBuilder.fieldHandleBean(fieldHandleBean);
                 return this;
             }
 
             @Override
             public ObjectBuilder type(@NotNull Class<?> type) {
-                objectBuilder.type(type);
+                abmBuilder.type(type);
+                return this;
+            }
+        }
+
+        private static class ObjectsBuilderImpl implements ObjectsBuilder {
+
+            private final List<ObjectBuilder> objects = new ArrayList<>();
+
+            /**
+             *
+             * @since 2025. 6. 20.
+             * @version 0.8.0
+             * @author parkjunhong77@gmail.com
+             *
+             * @see open.commons.spring.web.authority.metadata.AuthorizedMetadataBuilder.ObjectsBuilder#build()
+             */
+            @Override
+            public List<AuthorizedObjectMetadata> build() {
+                return objects.stream().map(b -> b.build()).collect(Collectors.toList());
+            }
+
+            /**
+             *
+             * @since 2025. 6. 20.
+             * @version 0.8.0
+             * @author parkjunhong77@gmail.com
+             *
+             * @see open.commons.spring.web.authority.metadata.AuthorizedMetadataBuilder.ObjectsBuilder#object(java.util.function.Function)
+             */
+            @Override
+            public ObjectsBuilder object(Function<ObjectBuilder, ObjectBuilder> consumer) {
+                ObjectBuilder builder = new ObjectBuilderImpl();
+                objects.add(consumer.apply(builder));
                 return this;
             }
         }
@@ -356,7 +456,7 @@ public class AuthorizedMetadataBuilder {
         AuthorizedFieldMetadata build();
 
         /**
-         * {@link IUnauthorizedFieldHandler}를 구현한 {@link Bean} 이름을 설정합니다. <br>
+         * {@link AuthorizedField#fieldHandleBean()}에 해당하는 값을 설정합니다.
          * 
          * <pre>
          * [개정이력]
@@ -376,7 +476,7 @@ public class AuthorizedMetadataBuilder {
         FieldBuilder fieldHandleBean(String fieldHandleBean);
 
         /**
-         * 데이터 처리 유형을 설정합니다.<br>
+         * {@link AuthorizedField#handleType()}에 해당하는 값을 설정합니다.<br>
          * 
          * <pre>
          * [개정이력]
@@ -396,7 +496,7 @@ public class AuthorizedMetadataBuilder {
         FieldBuilder handleType(int handleType);
 
         /**
-         * {@link Field} 이름을 설정합니다. <br>
+         * {@link AuthorizedField#name()}에 해당하는 값을 설정합니다.<br>
          * 
          * <pre>
          * [개정이력]
@@ -478,7 +578,7 @@ public class AuthorizedMetadataBuilder {
         ObjectBuilder field(Function<FieldBuilder, FieldBuilder> consumer);
 
         /**
-         * {@link IUnauthorizedFieldHandler}를 구현한 {@link Bean} 이름을 설정합니다. <br>
+         * {@link AuthorizedObject#fieldHandleBean()}에 해당하는 값을 설정합니다.
          * 
          * <pre>
          * [개정이력]
@@ -498,7 +598,7 @@ public class AuthorizedMetadataBuilder {
         ObjectBuilder fieldHandleBean(String fieldHandleBean);
 
         /**
-         * 권한설정 데이터 유형을 설정합니다. <br>
+         * {@link AuthorizedObject}가 적용되는 데이터 유형을 설정합니다.<br>
          * 
          * <pre>
          * [개정이력]
@@ -508,6 +608,7 @@ public class AuthorizedMetadataBuilder {
          * </pre>
          *
          * @param type
+         *            데이터 유형
          * @return
          *
          * @since 2025. 6. 18.
@@ -515,5 +616,12 @@ public class AuthorizedMetadataBuilder {
          * @author Park, Jun-Hong parkjunhong77@gmail.com
          */
         ObjectBuilder type(@NotNull Class<?> type);
+    }
+
+    public interface ObjectsBuilder {
+
+        List<AuthorizedObjectMetadata> build();
+
+        ObjectsBuilder object(Function<ObjectBuilder, ObjectBuilder> consumer);
     }
 }